move from comp to page and make language dropdown
